Moved new filters into LookupFilters
WIP: migrating to new filtering
