+++ SPID/cmake/headerlist.cmake
@@ -4,7 +4,6 @@ set(headers ${headers}
 	include/Distribute.h
 	include/DistributeManager.h
 	include/DistributePCLevelMult.h
-	include/Filters.h
 	include/FormData.h
 	include/KeywordDependencies.h
 	include/LogBuffer.h
